+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- `NoSuchPathError` when creating a repo without a mapping specified.
+
 ## [0.4.0] - 2022-04-12
 ### Added
 - Support for [Yarn](https://github.com/FabricMC/yarn) mappings.